project structure correct
