+++ goals/models/dates_model_mixin.py
@@ -1,0 +1,22 @@
+from django.db import models
+from django.utils import timezone
+
+
+# ----------------------------------------------------------------
+# custom mixin
+class DatesModelMixin(models.Model):
+    created = models.DateTimeField(
+        verbose_name="Дата создания"
+    )
+    updated = models.DateTimeField(
+        verbose_name="Дата последнего обновления"
+    )
+
+    def save(self, *args, **kwargs):
+        if not self.id:
+            self.created = timezone.now()
+        self.updated = timezone.now()
+        return super().save(*args, **kwargs)
+
+    class Meta:
+        abstract = True
