+++ goals/permissions.py
@@ -1,7 +1,9 @@
 from rest_framework import permissions
 
-from goals import models
-from goals.models import BoardParticipant
+from goals.models.board import BoardParticipant
+from goals.models.category import GoalCategory
+from goals.models.comment import Comment
+from goals.models.goal import Goal
 
 
 # ----------------------------------------------------------------
@@ -26,17 +28,17 @@ class CategoryPermissions(permissions.BasePermission):
     def has_object_permission(self, request, view, obj) -> bool:
         """Method to check category permissions"""
         if request.method in permissions.SAFE_METHODS:
-            return models.GoalCategory.objects.select_related('board').filter(
+            return GoalCategory.objects.select_related('board').filter(
                 board__participants__user=request.user,
                 board=obj.board
             ).exists()
         else:
-            return models.GoalCategory.objects.select_related('board').filter(
+            return GoalCategory.objects.select_related('board').filter(
                 board__participants__user=request.user,
                 board=obj.board,
                 board__participants__role__in=[
-                    models.BoardParticipant.Role.owner,
-                    models.BoardParticipant.Role.writer
+                    BoardParticipant.Role.owner,
+                    BoardParticipant.Role.writer
                 ]
             ).exists()
 
@@ -47,17 +49,17 @@ class GoalPermissions(permissions.BasePermission):
     def has_object_permission(self, request, view, obj) -> bool:
         """Method to check goal permissions"""
         if request.method in permissions.SAFE_METHODS:
-            return models.Goal.objects.select_related('category').filter(
+            return Goal.objects.select_related('category').filter(
                 category__board__participants__user=request.user,
                 category__board=obj.category.board
             ).exists()
         else:
-            return models.Goal.objects.select_related('category').filter(
+            return Goal.objects.select_related('category').filter(
                 category__board__participants__user=request.user,
                 category__board=obj.category.board,
                 category__board__participants__role__in=[
-                    models.BoardParticipant.Role.owner,
-                    models.BoardParticipant.Role.writer
+                    BoardParticipant.Role.owner,
+                    BoardParticipant.Role.writer
                 ]
             ).exists()
 
@@ -68,16 +70,16 @@ class CommentPermissions(permissions.BasePermission):
     def has_object_permission(self, request, view, obj) -> bool:
         """Method to check comment permissions"""
         if request.method in permissions.SAFE_METHODS:
-            return models.Comment.objects.select_related('goal').filter(
+            return Comment.objects.select_related('goal').filter(
                 goal__category__board__participants__user=request.user,
                 goal__category__board=obj.goal.category.board
             ).exists()
         else:
-            return models.Comment.objects.select_related('goal').filter(
+            return Comment.objects.select_related('goal').filter(
                 goal__category__board__participants__user=request.user,
                 goal__category__board=obj.goal.category.board,
                 goal__category__board__participants__role__in=[
-                    models.BoardParticipant.Role.owner,
-                    models.BoardParticipant.Role.writer
+                    BoardParticipant.Role.owner,
+                    BoardParticipant.Role.writer
                 ]
             ).exists()
