+++ goals/urls.py
@@ -1,21 +1,23 @@
 from django.urls import path
 
-from goals import views
-
+from goals.views.board import BoardCreateView, BoardListView, BoardDetailView
+from goals.views.category import CategoryCreateView, CategoryListView, CategoryDetailView
+from goals.views.comment import CommentCreateView, CommentListView, CommentDetailView
+from goals.views.goal import GoalCreateView, GoalListView, GoalDetailView
 
 # ----------------------------------------------------------------
 # urlpatterns
 urlpatterns = [
-    path("goal_category/create", views.CategoryCreateView.as_view()),
-    path("goal_category/list", views.CategoryListView.as_view()),
-    path("goal_category/<int:pk>", views.CategoryRetrieveUpdateDestroyView.as_view()),
-    path('goal/create', views.GoalCreateView.as_view()),
-    path('goal/list', views.GoalListView.as_view()),
-    path('goal/<int:pk>', views.GoalRetrieveUpdateDestroyView.as_view()),
-    path('goal_comment/create', views.CommentCreateView.as_view()),
-    path('goal_comment/list', views.CommentListView.as_view()),
-    path('goal_comment/<int:pk>', views.CommentRetrieveUpdateDestroyView.as_view()),
-    path('board/create', views.BoardCreateView.as_view()),
-    path('board/list', views.BoardListView.as_view()),
-    path('board/<int:pk>', views.BoardRetrieveUpdateDestroyView.as_view())
+    path('board/create', BoardCreateView.as_view(), name='board-create'),
+    path('board/list', BoardListView.as_view(), name='board-list'),
+    path('board/<int:pk>', BoardDetailView.as_view(), name='board-detail'),
+    path("goal_category/create", CategoryCreateView.as_view(), name="category-create"),
+    path("goal_category/list", CategoryListView.as_view(), name="category-list"),
+    path("goal_category/<int:pk>", CategoryDetailView.as_view(), name="category-detail"),
+    path('goal/create', GoalCreateView.as_view(), name="goal-create"),
+    path('goal/list', GoalListView.as_view(), name='goal-list'),
+    path('goal/<int:pk>', GoalDetailView.as_view(), name='goal-detail'),
+    path('goal_comment/create', CommentCreateView.as_view(), name='comment-create'),
+    path('goal_comment/list', CommentListView.as_view(), name='comment-list'),
+    path('goal_comment/<int:pk>', CommentDetailView.as_view(), name='comment-detail'),
 ]
